feat: refactor HomeUnregistered structure and implement map touch handling
