+++ MOBILE/app/src/main/java/com/project/mobile/fragments/Unregistered/HomeUnregistered.java
@@ -1,4 +1,4 @@
-package com.project.mobile.fragments;
+package com.project.mobile.fragments.Unregistered;
 
 import android.os.Bundle;
