+++ MOBILE/app/src/main/res/layout/fragment_home_unregistered.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.HomeUnregistered">
+    tools:context=".fragments.Unregistered.HomeUnregistered">
     <FrameLayout
         android:id="@+id/search_container"
         android:layout_width="match_parent"
@@ -12,14 +12,7 @@
         android:elevation="8dp"
 
         />
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-        <FrameLayout
-            android:id="@+id/map_container"
-            android:layout_width="match_parent"
-            android:layout_height="500dp" />
+
         <ScrollView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
@@ -29,7 +22,10 @@
                 android:layout_height="wrap_content"
                 android:orientation="vertical"
                 >
-
+                <FrameLayout
+                    android:id="@+id/map_container"
+                    android:layout_width="match_parent"
+                    android:layout_height="500dp" />
                 <LinearLayout
                     android:orientation="vertical"
                     android:layout_width="match_parent"
@@ -82,6 +78,6 @@
             </LinearLayout>
         </ScrollView>
 
-    </LinearLayout>
+
 
 </FrameLayout>
